Fixed issue with building i386 objs Still can't build x86_64

+++ libical/scripts/build.sh
@@ -10,38 +10,44 @@ SCRIPTS_DIR=$WORKING_DIR/../scripts
 echo "WORKING_DIR : $WORKING_DIR"
 echo "SCRIPTS_DIR : $SCRIPTS_DIR"
 
+LIPO="xcrun -sdk iphoneos lipo"
+
 # Download the lastest build Tools
-curl -OL http://ftpmirror.gnu.org/autoconf/autoconf-2.68.tar.gz
-tar xzf autoconf-2.68.tar.gz
-cd  autoconf-2.68
-./configure --prefix=$WORKING_DIR/autotools-bin
-make
-make install
-export PATH=$WORKING_DIR/autotools-bin/bin:$PATH
-cd $WORKING_DIR
-
-curl -OL http://ftpmirror.gnu.org/automake/automake-2.9.tar.gz
-tar xzf automake-2.9.tar.gz
-cd automake-2.9
-./configure --prefix=$WORKING_DIR/autotools-bin
-make
-make install
-cd $WORKING_DIR
-
-curl -OL http://ftpmirror.gnu.org/libtool/libtool-2.4.2.tar.gz
-tar xzf libtool-2.4.2.tar.gz
-cd libtool-2.4.2
-./configure --prefix=$WORKING_DIR/autotools-bin
-make
-make install
-cd $WORKING_DIR
-
+#curl -OL http://ftpmirror.gnu.org/autoconf/autoconf-2.68.tar.gz
+#tar xzf autoconf-2.68.tar.gz
+#cd  autoconf-2.68
+#./configure --prefix=$WORKING_DIR/autotools-bin
+#make
+#make install
+#export PATH=$WORKING_DIR/autotools-bin/bin:$PATH
+#cd $WORKING_DIR
+#
+#curl -OL http://ftpmirror.gnu.org/automake/automake-1.13.4.tar.gz
+#tar xzf automake-1.13.4.tar.gz
+#cd automake-1.13.4
+#./configure --prefix=$WORKING_DIR/autotools-bin
+#make
+#make install
+#cd $WORKING_DIR
+#
+#curl -OL http://ftpmirror.gnu.org/libtool/libtool-2.4.2.tar.gz
+#tar xzf libtool-2.4.2.tar.gz
+#cd libtool-2.4.2
+#./configure --prefix=$WORKING_DIR/autotools-bin
+#make
+#make install
+#cd $WORKING_DIR
+#
 # Download the latest library
-LIBRARY_DISTRO_URL="http://downloads.sourceforge.net/project/freeassociation/libical/libical-1.0/libical-1.0.tar.gz"
-
 LIBRARY_TARBALL="libical.tar.gz"
-curl -L  $LIBRARY_DISTRO_URL -o $LIBRARY_TARBALL 
-gunzip -c $LIBRARY_TARBALL| tar xopf -
+if [ ! -e  ./$LIBRARY_TARBALL ]; then
+  LIBRARY_DISTRO_URL="http://downloads.sourceforge.net/project/freeassociation/libical/libical-1.0/libical-1.0.tar.gz"
+
+
+  curl -L  $LIBRARY_DISTRO_URL -o $LIBRARY_TARBALL
+  gunzip -c $LIBRARY_TARBALL| tar xopf -
+fi
+
 LIBRARY_DIR="libical-1.0"
 pushd $LIBRARY_DIR > /dev/null
 export OUTPUTDIR="./lib"
@@ -55,17 +61,17 @@ ARCH=i386 $SCRIPTS_DIR/libical_make.sh
 ARCH=x86_64 $SCRIPTS_DIR/libical_make.sh
 
 # build up fat library
-lipo \
+$LIPO \
     -arch arm64  $OUTPUTDIR/arm64/libical.a \
     -arch armv7  $OUTPUTDIR/armv7/libical.a \
     -arch armv7s $OUTPUTDIR/armv7s/libical.a \
-    -arch x86_64 $OUTPUTDIR/x86_64/libical-static.a \
+    -arch i386 $OUTPUTDIR/i386/libical.a \
     -create -output $OUTPUTDIR/libical.a
 
-# -arch i386 $OUTPUTDIR/i386/libical-static.a \
+#  x86_64 is failing in the configuration step
+#    -arch x86_64 $OUTPUTDIR/x86_64/libical-static.a \
 
 cp  $OUTPUTDIR/libical.a ../../lib
-
 cp  ./src/libical/ical.h ../../src/include
 
 # Return to previous working directory
